+++ js/main.js
@@ -1,10 +1,15 @@
 /**
- * @author:
- * @date:
+ * Module Description
+ *
+ * @author Author Name
+ * @date 2013-01-01
  */
 
 require([
-	// Require the application
-], function () {
+	// Require the modules
+	'modules/module'
+], function (module) {
 	'use strict';
+
+	console.log(module);
 });